fix filters type
